# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Store API
  version: 1.0.0
  description: API para gerenciamento de produtos e pedidos

paths:
  /api/store/product/{id}:
    get:
      summary: Buscar produto por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    delete:
      summary: Deletar produto por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Produto deletado

  /api/store/product/get-all:
    get:
      summary: Listar todos os produtos
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

  /api/store/product/update/{id}:
    put:
      summary: Editar produto
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Produto atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

  /api/store/product/save:
    post:
      summary: Criar novo produto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

  /api/store/order/{id}:
    get:
      summary: Buscar pedido por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    delete:
      summary: Deletar pedido por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Pedido deletado

  /api/store/order/get-all:
    get:
      summary: Listar todos os pedidos
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

  /api/store/order/update/{id}:
    put:
      summary: Atualizar pedido
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Pedido atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /api/store/order/save:
    post:
      summary: Criar novo pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Pedido criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        preco:
          type: number
          format: double
        descricao:
          type: string

    Order:
      type: object
      properties:
        id:
          type: integer
        endereco:
          type: string
        dataPedido:
          type: string
          format: date-time
        status:
          type: string
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'